♻️ refactor: 모달 테스트 Modal.stories로 이동 및 수정
